GH-56 - merge in changes. Not sure why it doesn't build locally

+++ pyrsia_node/src/block_chain/block.rs
@@ -20,7 +20,7 @@ pub struct Block {
     pub id: u64,
     pub hash: String,
     pub previous_hash: String,
-    pub timestamp: i64,
+    pub timestamp: u128,
     pub data: String,
     pub nonce: u64,
 }
